Empece a usar userdata como deberia usarse, con miembros estaticos!!

+++ Tp_final/Tp_final/DRAW.cpp
@@ -60,28 +60,28 @@ DRAW::~DRAW()
 {
 }
 
-Obj_Graf_Player* DRAW::createObjGraf(double ID, PLAYER_TYPE type)
+Obj_Graf_Player* DRAW::createObjGraf(unsigned int ID, PLAYER_TYPE type)
 {
 	Obj_Graf_Player *obj = new Obj_Graf_Player(ID, type);
 	this->mapObjGraf[ID] = obj;
 	return obj;
 }
 
-Obj_Graf_Enemy* DRAW::createObjGraf(double ID, ENEMY_TYPE type)
+Obj_Graf_Enemy* DRAW::createObjGraf(unsigned int ID, ENEMY_TYPE type)
 {
 	Obj_Graf_Enemy *obj = new Obj_Graf_Enemy(ID, type);
 	this->mapObjGraf[ID] = obj;
 	return obj;
 }
 
-Obj_Graf_Projectile* DRAW::createObjGraf(double ID, PROYECTILE_TYPE type)
+Obj_Graf_Projectile* DRAW::createObjGraf(unsigned int ID, PROYECTILE_TYPE type)
 {
 	Obj_Graf_Projectile *obj = new Obj_Graf_Projectile(ID, type);
 	this->mapObjGraf[ID] = obj;
 	return obj;
 }
 
-Obj_Graf_Ball* DRAW::createObjGraf(double ID, BALL_TYPE type)
+Obj_Graf_Ball* DRAW::createObjGraf(unsigned int ID, BALL_TYPE type)
 {
 	Obj_Graf_Ball *obj = new Obj_Graf_Ball(ID, type);
 	this->mapObjGraf[ID] = obj;
@@ -92,29 +92,29 @@ void DRAW::draw()
 {
 	this->drawLevel();
 
-	vector<double> keys;
+	vector<unsigned int> keys;
 
-	for (map<double, Obj_Graf*>::iterator it = this->mapObjGraf.begin(); it != this->mapObjGraf.end(); ++it)		// creo un vector con todas las keys del mapa de O. graficos
+	for (map<unsigned int, Obj_Graf*>::iterator it = this->mapObjGraf.begin(); it != this->mapObjGraf.end(); ++it)		// creo un vector con todas las keys del mapa de O. graficos
 		keys.push_back(it->first);
 
-	for (vector<double>::iterator it = keys.begin(); it < keys.end(); it++)			// barro todos los elementos del mapa
+	for (vector<unsigned int>::iterator it = keys.begin(); it < keys.end(); it++)			// barro todos los elementos del mapa
 		if (this->mapObjGraf[*it]->isActive())										// si el objeto grafico esta activo
 			this->mapObjGraf[*it]->draw();											// se dibuja
 }
 
-void DRAW::destroyObj(double ID)
+void DRAW::destroyObj(unsigned int ID)
 {
 	this->mapObjGraf[ID]->destroy();
 }
 
 void DRAW::destroyAll()
 {
-	vector<double> keys;
+	vector<unsigned int> keys;
 
-	for (map<double, Obj_Graf*>::iterator it = this->mapObjGraf.begin(); it != this->mapObjGraf.end(); ++it)		// creo un vector con todas las keys del mapa de O. graficos
+	for (map<unsigned int, Obj_Graf*>::iterator it = this->mapObjGraf.begin(); it != this->mapObjGraf.end(); ++it)		// creo un vector con todas las keys del mapa de O. graficos
 		keys.push_back(it->first);
 
-	for (vector<double>::iterator it = keys.begin(); it < keys.end(); it++)			// barro todos los elementos del mapa
+	for (vector<unsigned int>::iterator it = keys.begin(); it < keys.end(); it++)			// barro todos los elementos del mapa
 		this->mapObjGraf[*it]->destroy();										// si el objeto grafico esta activo
 }
 
@@ -123,7 +123,7 @@ void DRAW::setLevel(unsigned int level)
 	this->level = level;
 }
 
-bool DRAW::secuenceOver(double ID)
+bool DRAW::secuenceOver(unsigned int ID)
 {
 	return mapObjGraf[ID]->secuenceOver();
 }
